scripts/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
sentence-transformers>=3.0
supabase>=2.0
python-dotenv>=1.0
tiktoken>=0.7
fastapi>=0.115
uvicorn[standard]>=0.30
openai>=1.0
python-telegram-bot>=21.0
